fix c version types

+++ cfibo.c
@@ -3,8 +3,9 @@
 
 #include <stdlib.h>
 #include <stdio.h>
+#include <inttypes.h>
 
-int fib(int n) {
+int64_t fib(int64_t n) {
   if(n <= 0)
     return 0;
   if(n == 1)
@@ -19,7 +20,7 @@ int main(int argc, char const *argv[]) {
   }
 
   int n = atoi(argv[1]);
-  printf("%i\n", fib(n));
+  printf("%" PRIu64 "\n", fib(n));
 
   return 0;
 }
